requirements.txt
flask
flask-mysqldb
werkzeug
python-dotenv
requests
pandas
numpy
scikit-learn
